state: Move batch hunk cache out of navigation
Batch file rendering already lives in focused data modules, while hunk
navigation still owns the selected-state cache for rendered batch files.

The project lacks a focused owner for batch selected-state caching. Keeping
that persistence path in hunk navigation makes the navigation module
responsible for batch display storage decisions.

This commit addresses that by moving the rendered batch cache routines into
data.selected_change.batch_file_cache, redirecting the show-from and interactive callers,
and dropping the unused batch helper exports from hunk_tracking.

The next commit will add path-pure architecture coverage for the boundary.
